scripts/requirements.txt
requests>=2.31.0
pyyaml>=6.0
markdownify>=0.13.0
beautifulsoup4>=4.12.0
tenacity>=8.2.0
